[DELETE] PeopleInfo 내용 View 안으로 (#514)
